fix : 에러 핸들링을 하던 계층을 controller => service로 변경 및 esLint 패키지 설치

+++ services/posts.service.js
@@ -1,5 +1,8 @@
 const postRepository = require("../repositories/posts.repository");
-let arry2 = [];
+const { Hashtag } = require("../models");
+const log = require("../winston");
+let BadRequestError = require("./http-errors").BadRequestError;
+const help = require("korean-regexp");
 class PostService {
   postRepository = new postRepository();
   createPost = async (
@@ -13,6 +16,24 @@ class PostService {
     userId,
     nickname
   ) => {
+    if (!content) {
+      log.error("PostController.createPost : content is required");
+      throw new BadRequestError(
+        "PostController.createPost : content is required"
+      );
+    }
+    if (hashtag) {
+      let consonant = [];
+      //유저가 Hashtag를 입력했을 때 Hashtag table에 Hashtag를 생성
+      for (let i = 0; i < hashtag.length; i++) {
+        consonant[i] = help.explode(hashtag[i]).join("");
+        await Hashtag.create({
+          hashtag: hashtag[i],
+          consonant: consonant[i],
+          postId: createPost.postId,
+        });
+      }
+    }
     const createPost = await this.postRepository.createPost(
       content,
       time,
@@ -28,6 +49,12 @@ class PostService {
   };
 
   getAllPosts = async (pagenum, userId) => {
+    if (!pagenum) {
+      log.error("PostController.getAllPosts : pagenum is required");
+      throw new BadRequestError(
+        "PostController.getAllPosts : pagenum is required"
+      );
+    }
     const getAllPosts = await this.postRepository.getAllPosts(pagenum);
 
     return Promise.all(
@@ -39,6 +66,12 @@ class PostService {
     );
   };
   geLikeAllPosts = async (pagenum, userId) => {
+    if (!pagenum) {
+      log.error("PostController.getLikeAllPosts : pagenum is required");
+      throw new BadRequestError(
+        "PostController.getLikeAllPosts : pagenum is required"
+      );
+    }
     const getLikeAllPosts = await this.postRepository.getLikeAllPosts(pagenum);
 
     return Promise.all(
@@ -50,6 +83,10 @@ class PostService {
     );
   };
   getPost = async (postId, userId) => {
+    if (!postId) {
+      log.error("PostController.getPost : postId is required");
+      throw new BadRequestError("PostController.getPost : postId is required");
+    }
     const getPost = await this.postRepository.getPost(postId, userId);
 
     return getPost;
@@ -62,9 +99,31 @@ class PostService {
     path,
     speed,
     image,
-    hashtag,
-    checkHash
+    hashtag
   ) => {
+    let arrayHash = [];
+    let checkHash = false;
+    const checkSameHashTag = await Hashtag.findAll({
+      where: { postId },
+      attributes: ["hashtag"],
+    });
+
+    //유저가 게시글을 수정할 때 hashtag를 수정하면 Hashtag 테이블의 hashtag도 같이 수정
+    for (let i = 0; i < checkSameHashTag.length; i++) {
+      arrayHash.push(checkSameHashTag[i].hashtag);
+    }
+
+    if (hashtag.join("") === arrayHash.join("")) {
+      checkHash = true;
+    } else {
+      checkHash = false;
+    }
+    if (!postId) {
+      log.error("PostController.updatePost : postId is required");
+      throw new BadRequestError(
+        "PostController.updatePost : update is required"
+      );
+    }
     const updatePost = await this.postRepository.updatePost(
       postId,
       content,
@@ -79,14 +138,32 @@ class PostService {
     return updatePost;
   };
   deletePost = async (postId) => {
+    if (!postId) {
+      log.error("PostController.deletePost : postId is required");
+      throw new BadRequestError(
+        "PostController.deletePost : update is required"
+      );
+    }
     const deletePost = await this.postRepository.deletePost(postId);
     return deletePost;
   };
   searchPost = async (hashtag) => {
+    if (!hashtag) {
+      log.error("PostController.searchPost : hashtag is required");
+      throw new BadRequestError(
+        "PostController.searchPost : hashtag is required"
+      );
+    }
     const searchPost = await this.postRepository.searchPost(hashtag);
     return searchPost;
   };
   autoSearchPost = async (hashtag) => {
+    if (!hashtag) {
+      log.error("PostController.autoSearchPost : hashtag is required");
+      throw new BadRequestError(
+        "PostController.autoSearchPost : hashtag is required"
+      );
+    }
     const autoSearchPost = await this.postRepository.autoSearchPost(hashtag);
     return autoSearchPost;
   };
